MyInterpreter-v1.2/requirements.txt
mysql-connector-python
pytest
PyYAML
psycopg2-binary
numpy
cx_Oracle
pyodbc


